fix compile error

+++ src/rest.rs
@@ -86,7 +86,7 @@ impl BlockValue {
         BlockValue {
             id: header.block_hash().to_hex(),
             height: blockhm.header_entry.height() as u32,
-            version: header.version,
+            version: header.version as u32,
             timestamp: header.time,
             tx_count: blockhm.meta.tx_count,
             size: blockhm.meta.size,
@@ -152,7 +152,7 @@ impl TransactionValue {
 
         TransactionValue {
             txid: tx.txid(),
-            version: tx.version,
+            version: tx.version as u32,
             locktime: tx.lock_time,
             vin: vins,
             vout: vouts,
